rules: Sanitize value, unroll loop, add 5 new properties
The new properties are:

 * 'centered'
 * 'slave'
 * 'master'
 * 'tag'
 * 'new_tag'

The unrolling is done to improve performance as each check is performed
once rather than for every properties.

The additional code handle tag screen corder case and multiple type()
for each property. For example, a new tag can be a boolean "I want a
new tag", a tag (then it is the same as 'tag' or an array of tag
properties.