naughty: Fix builds without dbus
The new notification action code tried to emit dbus signals even if awesome was
build without notification support. Fix this by adding the necessary "if".

(This also removes an unused return value which wasn't marked as "local" and
thus triggered my "complain if something messes with the global env"-script)

Signed-off-by: Uli Schlachter <psychon@znc.in>