Revert "layout.arrange only once per main loop"
This reverts 7410646, 1777605 and c88c51f.

The problem is that the change was meant to improve performance, but
triggers problems where a plugin / config relies on geometries being
updated/changed during the same main loop.

The code might be improved to e.g. not delay arranging sometimes, but
then it's easier to restore the old behavior for now.

We might re-add this, but then should provide a way to easily call
something on the next arrange signal, e.g.

    > screen:on_next("arrange", function() ... end)