cmake: Really add '--dirty' flag to 'git describe' for version
This fixes commit 2dc14b10de9.

That commit added the '--dirty' argument the first time we generate the version
stamp file, which is when CMake produces the Makefiles (called from
awesomeConfig.cmake). Later on, we use build-utils/git-version-stamp.sh to
update this file. That file was missing this new argument to 'dit describe'

Signed-off-by: Uli Schlachter <psychon@znc.in>