config: sloppy focus: ignore mouse::enter if mouse was not moved
When raising a client's transient_for window, the parent window gets
also raised.  This might trigger a `mouse::enter` signal for the parent
window, if it was not visible at the mouse pointer's position before.

The sloppy focus handler in the default config now ignores any
`mouse::enter` events, when the mouse coordinates have not changed since
the last signal.