Fix mouse::leave for drawables (titlebar)
Store X/Y of the last motion event in `globalconf`, to have it available
in `event_handle_leavenotify`, where only the current X/Y is available
on the event.  And since the titlebar has only a drawable, but no
drawin, it will not get handled through `drawin_getbywin(ev->event)`.

An alternative might be to use a `drawin` also for the titlebar.