Apply left/top titlebars in event_handle_configurerequest
I have noticed that the HeidiSQL session manager window/dialog will be
moved down by the height of the top titlebar when selecting an entry
from the server list.

While I am not sure if that is the correct way to fix this, and what it
might break, it fixes the issue described above.