Add wibox.hierarchy:update()
This function updates a hierarchy if necessary and does nothing on the parts
that did not change. This should be more efficient than recomputing the whole
state whenever it is needed. :-)

The assumption here is that the result of a widget's callback can be compared
with ==. This means that a widget must never modify any table that it once
returned from its :layout callback. This commit also updates the docs to make
this requirement more explicit.

Signed-off-by: Uli Schlachter <psychon@znc.in>