wibox.drawable: Use hierarchy:update()
Instead of throwing away the whole widget hierarchy at every relayout, this now
just updates the hierarchy when needed. This avoids recomputing things that
didn't change

The results for the "benchmark test" before this patch:

        create wibox: 0.0833272  sec/iter ( 13 iters, 1.166 sec for benchmark)
    update textclock: 0.0191503  sec/iter ( 57 iters, 2.501 sec for benchmark)
  relayout textclock: 0.0157272  sec/iter ( 65 iters, 1.038 sec for benchmark)
    redraw textclock: 0.00165183 sec/iter (643 iters, 2.873 sec for benchmark)

After this patch, I get these results:

        create wibox: 0.0825847  sec/iter ( 13 iters, 1.155 sec for benchmark)
    update textclock: 0.00387662 sec/iter (260 iters, 3.014 sec for benchmark)
  relayout textclock: 0.0023906  sec/iter (420 iters, 1.006 sec for benchmark)
    redraw textclock: 0.00172511 sec/iter (634 iters, 2.005 sec for benchmark)

It can be seen that the relayout only takes about 15% of the time that it
previously needed. Similarly, a full update only takes 20% of the time that was
previously needed. Finally, the update takes half as much time as it did with
the old widget layouts (0.00736755 sec/iter).

Signed-off-by: Uli Schlachter <psychon@znc.in>