wibox.hierarchy:update(): Handle removed children
Imagine closing the last client on a tag. Afterwards the tasklist will be empty.
Thus, we need to redraw the tasklist to fill it with the background color.

This commit does exactly that: When some widget has less children than before
doing a relayout, the area covered by the children that were removed is added to
the dirty area.

Signed-off-by: Uli Schlachter <psychon@znc.in>