Fix an oversight in widget dependency tracking
A widget can position another widget without consulting its fit function for a
desired size. However, the new hierarchy:update() function assumes that nothing
changed in this sub-hierarchy if layout_widget() returns the same result and
thus this missed changes.

Fix this by adding a dependency between widgets when one widget places another
widget. This fixes e.g. the tasklist not updating when the order of two clients
was swapped.

Signed-off-by: Uli Schlachter <psychon@znc.in>